Раскиданы переменные, обновлены функции
