--- src/config.js
@@ -1,10 +1,0 @@
-const POSTERS_PATH = `./images/posters/`;
-const KEY_CODE = {
-  ESC: `Escape`,
-  ENTER: `Enter`
-};
-
-export {
-  POSTERS_PATH,
-  KEY_CODE,
-};
